Link search oj bucketing to mapped variants [copilot]
